Split reusable RepoManager into a shared impl base class
